feat: extract gateway frontend package workspace
