+++ package.json
@@ -2,6 +2,10 @@
   "name": "twofactor_gateway",
   "version": "3.0.0-dev.0",
   "description": "A two-factor auth provider for Nextcloud",
+  "private": true,
+  "workspaces": [
+    "packages/twofactor-gateway"
+  ],
   "type": "module",
   "scripts": {
     "build": "vite build --mode production",
@@ -23,6 +27,7 @@
   "author": "Christoph Wurst <christoph@winzerhof-wurst.at>",
   "license": "AGPL-3.0",
   "dependencies": {
+	"@libresign/twofactor-gateway": "workspace:*",
     "@nextcloud/axios": "^2.6.0",
     "@nextcloud/initial-state": "^3.0.0",
     "@nextcloud/l10n": "^3.4.1",
