+++ packages/twofactor-gateway/README.md
@@ -1,0 +1,10 @@
+# @libresign/twofactor-gateway
+
+Internal reusable frontend package for TwoFactor Gateway and LibreSign integrations.
+
+It currently provides:
+- Shared gateway types
+- Admin gateway service layer
+- View-model helpers for AdminSettings
+- Gateway instance modal domain model
+- Basic composables for gateway API/form state
