+++ packages/twofactor-gateway/package.json
@@ -1,0 +1,16 @@
+{
+  "name": "@libresign/twofactor-gateway",
+  "version": "1.0.0",
+  "private": true,
+  "type": "module",
+  "main": "./src/index.ts",
+  "types": "./src/index.ts",
+  "exports": {
+    ".": "./src/index.ts"
+  },
+  "peerDependencies": {
+    "@nextcloud/axios": "^2.6.0",
+    "@nextcloud/router": "^3.1.0",
+    "vue": "^3.5.0"
+  }
+}
