+++ packages/twofactor-gateway/src/composables/index.ts
@@ -1,0 +1,2 @@
+export * from './useGatewayApi.ts'
+export * from './useGatewayForm.ts'
