+++ vitest.config.js
@@ -3,8 +3,14 @@
 
 import { mergeConfig } from 'vitest/config'
 import vue from '@vitejs/plugin-vue'
+import { fileURLToPath, URL } from 'node:url'
 
 export default mergeConfig({
+	resolve: {
+		alias: {
+			'@libresign/twofactor-gateway': fileURLToPath(new URL('./packages/twofactor-gateway/src/index.ts', import.meta.url)),
+		},
+	},
 	plugins: [vue()],
 }, {
 	test: {
